requirements.txt
discord_py_slash_command==2.3.2
discord.py==1.7.3
requests==2.25.1
Django==3.2.4
discord==1.7.3
python-dotenv==0.19.0
